requirements.txt
Alfred-Workflow==1.17.2
requests==2.9.1
